+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - FIRMS provider
 - EM-DAT provider
 - Static data providers (tornado.canada-gov, tornado.australian-bm, tornado.osm-wiki, wildfire.frap.cal, tornado.des-inventar-sendai)
-- Noaa provider
+- storms.noaa provider
 #### Changed
 - Event versioning is removed 
 - Refactored normalization, event and episods jobs
@@ -22,7 +22,7 @@ emdat:
 konturApi:
   host: 'https://api.kontur.io/'
 
-noaaTornado:
+stormsNoaa:
   host: 'https://www1.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/'
 
 staticdata:
@@ -41,7 +41,7 @@ scheduler:
   staticImport:
     enable: true
     initialDelay: 1000
-  noaaTornadoImport:
+  stormsNoaaImport:
     enable: true
     initialDelay: 1000
     fixedDelay: P30D # every 30 days